adding a lot of dbflow to implement saving. It doesn't fully work now, but almost there

+++ app/src/main/java/com/example/ianshinbro/trackerbat/Implentation/iPlayer.java
@@ -16,7 +16,7 @@ public interface iPlayer {
     void setLastName(String lastName);
     void setNickName(String nickName);
 
-     int getNumber();
+     long getNumber();
 
      void setNumber(int number);
