+++ app/src/main/java/com/example/ianshinbro/trackerbat/UI/popupScreens/UpdateGame.java
@@ -101,7 +101,7 @@ public class UpdateGame extends Activity {
         updateGame.setText(R.string.SaveChanges_BTNText);
         homeTeam.setText(game.getHomeTeam());
         awayTeam.setText(game.getAwayTeam());
-        inningTotal.setText(Integer.toString(game.getInningTotal()));
+        inningTotal.setText(Long.toString(game.getInningTotal()));
     }
 
     /**
